backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
geopy==2.4.0
matplotlib==3.7.2
openai>=1.0.0
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
faiss-cpu==1.7.4
pytz==2023.3
python-docx==0.8.11
google-auth==2.33.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.141.0
gunicorn==21.2.0
